Fixed buggy mechanics and changed paste mechanics
Issues with normal, non-op users being able to place blueprints in claims, and also not in restricted/sub claims.

Also changed the way blueprints are blocked in the pasting process. Instead of one block in the entire region, only blocks that would be replaced by the paste will be considered "in the way".